requirements.txt
psycopg2==2.4.6
psycogreen==1.0
pyramid==1.4
pyramid_tm==0.7
sqlalchemy==0.8.0
gevent==0.13.8
gunicorn==0.17.2
zope.sqlalchemy==0.7.2
